Tests/LibWeb/Text/expected/HTML/shadow-realm-url.txt

Shadow realm evaluation returned: example.com
